Add methods to UserManager and ResourcesManager

+++ PAS/src/main/java/pl/pas/model/managers/ResourcesManager.java
@@ -1,14 +1,14 @@
 package pl.pas.model.managers;
 
+import pl.pas.model.entities.Borrow;
+import pl.pas.model.entities.resource.AudioBook;
 import pl.pas.model.entities.resource.Book;
 import pl.pas.model.entities.resource.Resource;
-import pl.pas.model.repositories.ResourceRepository;
 import pl.pas.model.repositories.interfaces.IBorrowRepository;
 import pl.pas.model.repositories.interfaces.IResourceRepository;
 import pl.pas.model.repositories.interfaces.IUserRepository;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 public class ResourcesManager {
     IResourceRepository resourceRepository;
@@ -32,21 +32,59 @@ public class ResourcesManager {
         return resourceRepository.getAllResources();
     }
 
+    public List<Resource> getAllBooks() {
+        ArrayList<Resource> books = new ArrayList<>();
+        for (Resource resource : resourceRepository.getAllResources()) {
+            if (resource instanceof Book) {
+                books.add(resource);
+            }
+        }
+        return books;
+    }
+
+    public List<Resource> getAllAudioBooks() {
+        ArrayList<Resource> audioBooks = new ArrayList<>();
+        for (Resource resource : resourceRepository.getAllResources()) {
+            if (resource instanceof AudioBook) {
+                audioBooks.add(resource);
+            }
+        }
+        return audioBooks;
+    }
+
     public boolean removeResource(Resource resource) {
-        return resourceRepository.deleteResource(resource.getResourceId());
+        return removeResource(resource.getResourceId());
     }
 
     public boolean removeResource(UUID uuid) {
+        if (uuid == null || resourceRepository.getResource(uuid) == null) {
+            return false;
+        }
+        if (!resourceRepository.getResource(uuid).isAvailable()) {
+            return false;
+        }
+        for (Borrow borrow : borrowRepository.getAllBorrows()) {
+            if (borrow.getResource() == resourceRepository.getResource(uuid)) {
+                borrow.setResource(null);
+            }
+        }
         return resourceRepository.deleteResource(uuid);
     }
 
-    public boolean createBook(long isbn, String title, String author, int publishYear) {
+    public boolean addBook(long isbn, String title, String author, int publishYear) {
+        if (isbn == 0 || title == null || author == null || publishYear > Calendar.getInstance().get(Calendar.YEAR)) {
+            return false;
+        }
         return resourceRepository.addResource(new Book(isbn, title, author, publishYear), UUID.randomUUID());
     }
 
-    public boolean createAudioBook(long isbn, String title, String author, int length) {
+    public boolean addAudioBook(long isbn, String title, String author, int length) {
+        if (isbn == 0 || title == null || author == null || length <= 0) {
+            return false;
+        }
         return resourceRepository.addResource(new Book(isbn, title, author, length), UUID.randomUUID());
     }
+
